JCLOUDS-847: Poor upload performance for putBlob
This change adds a BufferedOutputStream around the output stream used
when uploading blobs with the default Java URL connection HTTP client.

The buffer size is controlled by the following property with the
following default value:

jclouds.outputstream.buffer-size: 32768

The buffered output stream wraps the counting output stream and not vice
versa, so that counting occurs for bytes that are actually written or
flushed to the socket.

On some platforms this will provide not just better throughput, but also
lower CPU consumption.